backend/pyproject.toml
[tool.poetry]
name = "tokyo-guide-backend"
version = "0.1.0"
description = "Tokyo Travel Guide Backend - FastAPI with RAG chat and Telegram bot"
authors = ["blevy"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
groq = "^1.0.0"
sentence-transformers = "^3.0.0"
supabase = "^2.7.0"
python-telegram-bot = "^21.0"
beautifulsoup4 = "^4.12.0"
httpx = "^0.27.0"
pydantic-settings = "^2.5.0"
python-dotenv = "^1.0.0"
lxml = "^5.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^5.0.0"
ruff = "^0.6.0"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
